Refactor error handling in ImgFile extension
The error handling in the ImgFile extension has been refactored to improve clarity and order of operations. The condition checking if the filePath starts with "http://" or "https://" has been moved inside the guard statement for better readability. A case to return nil in URL initialization when filePath starts with "http://" or "https://" has also been added.