Add version.txt file to track build versions
A new file named version.txt has been added to the .gitignore and scripts. This file is intended to track the version of the software build. The build and clean scripts have been updated to interact with this new file. Package.swift has also been modified to include version.txt as a resource for relevant packages.