Add version input to Github Workflow release
In this update, we've added an optional 'version' input to the Github Workflow release. Users can now specify the version number during a manual release. If no version is provided, the script will use the latest tag as the version. Fixed an issue in the upload script where the release version was not properly passed to the Github release command.