Make release version input mandatory
The code has been revised to make the release version input a requirement in the GitHub workflow. This enhances the workflow by using the provided version for the release, instead of determining it with git. With this change, the scripts now also consider the command line input version and exit early with an error if no version is provided or if the release already exists.